+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2025-08-09
+
+- `$foo("a", "b")` and other messages starting with `$` are now treated as JavaScript code to execute.
+- ...except for stuff like `$foo a b c` which is treated as a command invocation.
+- `$foo` can now be used in JS instead of `$.cmd("foo")`.
+
 ## 2025-08-02
 
 - Added support for `$` prefix as an alternative to `.` for accessing aliases (e.g., `$weather` works the same as `.weather`).